snowpark_requirements.txt
snowflake-snowpark-python[pandas]
boto3